fixed issue in repository class

+++ hibernate_ManyToMany/src/main/java/lk/harshana/dao/StudentDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import lk.harshana.model.Student;
 
-public interface StudentDao extends CrudRepository<Student, Long> {
+public interface StudentDao extends CrudRepository<Student, Integer> {
 
 }
